Support new TF2 dtype behavior in Keras layers.
In TensorFlow 2, layers will default to float32 and cast inputs to the layer's dtype. This means many existing float64 models will be silently start using float32 in TF 2 (although a warning is issued).

This change makes code in tensorflow_probability/python/layers support this new behavior. In particular, whenever one layer creates another layer, the outer layer now passes the correct dtype to the inner layer.

It also fixes all tests in tensorflow_probability/python/layers to run in the same dtype as in TF 1. Doing this mostly consisted of passing "dtype=self.dtype" to layer constructors. I also could have set floatx to float64, which would have been simpler, but then I would not test that outer layers pass the correct dtype to inner layers.

In order to ensure the dtypes in the test are the same as in TF 1, I manually checked that the warning TF 2 issues when the new behavior is encountered does not appear. Some assertions checking dtypes are also added.

PiperOrigin-RevId: 261247104

+++ tensorflow_probability/python/layers/variable_input_test.py
@@ -37,10 +37,11 @@ class VariableInputLayerTest(tf.test.TestCase):
             dtype=tf.float64,
             trainable=False),  # You'd probably never want this in IRL.
         # The Dense serves no real purpose; it will change the event_shape.
-        tf.keras.layers.Dense(5, use_bias=False),
+        tf.keras.layers.Dense(5, use_bias=False, dtype=tf.float64),
         tfp.layers.DistributionLambda(
             lambda t: tfd.Independent(tfd.Normal(loc=t[0], scale=t[1]),  # pylint: disable=g-long-lambda
-                                      reinterpreted_batch_ndims=1)),
+                                      reinterpreted_batch_ndims=1),
+            dtype=tf.float64),
 
     ])
 
@@ -77,10 +78,11 @@ class VariableInputLayerTest(tf.test.TestCase):
         trainable=False,  # You'd probably never want this in IRL.
     )(dummy_input)
     # The Dense serves no real purpose; it will change the event_shape.
-    x = tf.keras.layers.Dense(5, use_bias=False)(x)
+    x = tf.keras.layers.Dense(5, use_bias=False, dtype=tf.float64)(x)
     x = tfp.layers.DistributionLambda(
         lambda t: tfd.Independent(tfd.Normal(loc=t[0], scale=t[1]),  # pylint: disable=g-long-lambda
-                                  reinterpreted_batch_ndims=1))(x)
+                                  reinterpreted_batch_ndims=1),
+        dtype=tf.float64)(x)
     model = tf.keras.Model(dummy_input, x)
 
     # Instantiate the model (as a TFP distribution).
